fix: レビュー指摘（HIGH/SHOULD）の修正 - セキュリティ・型安全性・リソース管理を改善
- Cookie Secureフラグを環境変数で制御（APP_ENV=production時のみtrue）
- GetCurrentUserの型をint64に統一（JWT ClaimのUserIDと型を合わせる）
- UNIQUE制約エラー検出をGORM標準のgorm.ErrDuplicatedKeyに変更
- レート制限のgoroutineにcontext.Contextを追加してリーク防止
- JWT_SECRETプレースホルダーを64文字以上に変更
- main.goのcontext変数名を明確化（shutdownCtx）

+++ backend/internal/services/auth_service.go
@@ -247,13 +247,13 @@ func (s *AuthService) Logout(userID int64) error {
 }
 
 // GetCurrentUser は現在のユーザー情報を取得
-func (s *AuthService) GetCurrentUser(userID int) (*models.User, error) {
+func (s *AuthService) GetCurrentUser(userID int64) (*models.User, error) {
 	logging.L.Debug("getting current user",
 		"service", "AuthService",
 		"method", "GetCurrentUser",
 		"user_id", userID)
 
-	user, err := s.userRepo.GetByID(userID)
+	user, err := s.userRepo.GetByID(int(userID))
 	if err != nil {
 		logging.L.Error("failed to get user",
 			"service", "AuthService",
